requirements.txt
uvicorn
fastapi
pytest
httpx
psycopg2-binary
dotenv